DEV-129 Change packages structure in Requests
