+++ src/main/java/com/efarm/efarmbackend/payload/request/auth/UpdateActivationCodeByLoggedOwnerRequest.java
@@ -1,4 +1,4 @@
-package com.efarm.efarmbackend.payload.request;
+package com.efarm.efarmbackend.payload.request.auth;
 
 import jakarta.validation.constraints.NotBlank;
 import lombok.Getter;
