+++ src/main/java/com/efarm/efarmbackend/service/landparcel/LandparcelFacade.java
@@ -4,8 +4,8 @@ import com.efarm.efarmbackend.model.farm.Farm;
 import com.efarm.efarmbackend.model.landparcel.Landparcel;
 import com.efarm.efarmbackend.model.landparcel.LandparcelDTO;
 import com.efarm.efarmbackend.model.landparcel.LandparcelId;
-import com.efarm.efarmbackend.payload.request.AddLandparcelRequest;
-import com.efarm.efarmbackend.payload.request.UpdateLandparcelRequest;
+import com.efarm.efarmbackend.payload.request.landparcel.AddLandparcelRequest;
+import com.efarm.efarmbackend.payload.request.landparcel.UpdateLandparcelRequest;
 import com.efarm.efarmbackend.repository.landparcel.LandparcelRepository;
 import com.efarm.efarmbackend.service.user.UserService;
 import jakarta.transaction.Transactional;
