+++ src/test/groovy/com/efarm/efarmbackend/service/equipment/FarmEquipmentFacadeSpec.groovy
@@ -1,7 +1,7 @@
 package com.efarm.efarmbackend.service.facades
 
 import com.efarm.efarmbackend.model.equipment.FarmEquipment
-import com.efarm.efarmbackend.payload.request.AddUpdateFarmEquipmentRequest
+import com.efarm.efarmbackend.payload.request.equipment.AddUpdateFarmEquipmentRequest
 import com.efarm.efarmbackend.model.equipment.FarmEquipmentId
 import com.efarm.efarmbackend.model.equipment.EquipmentCategory
 import com.efarm.efarmbackend.model.farm.Farm
