+++ src/test/groovy/com/efarm/efarmbackend/service/farm/FarmServiceSpec.groovy
@@ -6,7 +6,7 @@ import com.efarm.efarmbackend.model.user.User
 import com.efarm.efarmbackend.model.user.Role
 import com.efarm.efarmbackend.model.user.ERole
 import com.efarm.efarmbackend.repository.farm.ActivationCodeRepository
-import com.efarm.efarmbackend.payload.request.UpdateFarmDetailsRequest
+import com.efarm.efarmbackend.payload.request.farm.UpdateFarmDetailsRequest
 import com.efarm.efarmbackend.repository.farm.FarmRepository
 import com.efarm.efarmbackend.repository.user.UserRepository
 import com.efarm.efarmbackend.service.farm.FarmService
