+++ src/test/java/com/efarm/efarmbackend/controller/FarmEquipmentControllerIT.java
@@ -16,7 +16,7 @@ import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.MvcResult;
 
 import com.efarm.efarmbackend.model.equipment.FarmEquipment;
-import com.efarm.efarmbackend.payload.request.AddUpdateFarmEquipmentRequest;
+import com.efarm.efarmbackend.payload.request.equipment.AddUpdateFarmEquipmentRequest;
 import com.efarm.efarmbackend.model.equipment.FarmEquipmentId;
 import com.efarm.efarmbackend.model.farm.Farm;
 import com.efarm.efarmbackend.model.user.User;
